# api/openapi.json
{
  "components": {
    "schemas": {
...
      "PublicTrustCenterResponse": {
        "description": "Strictly allow-listed public trust center representation.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "headline": {
            "title": "Headline",
            "type": "string"
          },
          "organization_slug": {
            "title": "Organization Slug",
            "type": "string"
          },
          "overview": {
            "title": "Overview",
            "type": "string"
          },
          "primary_color": {
            "title": "Primary Color",
            "type": "string"
          },
          "resources": {
            "items": {
              "$ref": "#/components/schemas/TrustResourceResponse"
            },
            "title": "Resources",
            "type": "array"
          },
          "security_contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Security Contact Email"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "organization_slug",
          "display_name",
          "headline",
          "overview",
          "security_contact_email",
          "primary_color",
          "updated_at",
          "resources"
        ],
        "title": "PublicTrustCenterResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Watchtower GRC maintainers",
      "url": "https://github.com/guiltykeyboard/msp-vciso"
    },
    "description": "Watchtower's versioned HTTP API for MSP, customer, auditor, evidence, and assessment workflows. The X-Watchtower headers currently shown on tenant endpoints are an insecure development adapter, not production authentication.",
    "license": {
      "identifier": "AGPL-3.0-only",
      "name": "GNU Affero General Public License v3.0"
    },
    "summary": "Tenant-safe compliance and evidence automation for MSPs",
    "title": "Watchtower GRC API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/public/trust": {
      "get": {
        "description": "Resolve only explicitly published, public-safe data without tenant headers.",
        "operationId": "getPublicTrustCenter",
        "parameters": [
          {
            "in": "query",
            "name": "slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicTrustCenterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a published trust center by hostname or tenant slug",
        "tags": [
          "public trust centers"
        ]
      }
    },
...
    "/v1/trust-center": {
      "get": {
        "description": "Return tenant-managed public content, resources, and domain states.",
        "operationId": "getTrustCenter",
        "parameters": [
          {
            "in": "header",
            "name": "X-Watchtower-Organization",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Watchtower-Organization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Watchtower-Actor",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Watchtower-Actor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustCenterManagementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get tenant trust center configuration",
        "tags": [
          "trust centers"
        ]
      },
      "put": {
        "description": "Upsert the allow-listed profile fields and explicit publication state.",
        "operationId": "updateTrustCenter",
        "parameters": [
          {
            "in": "header",
            "name": "X-Watchtower-Organization",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Watchtower-Organization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Watchtower-Actor",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Watchtower-Actor",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrustCenterProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustCenterManagementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save and publish a tenant trust center",
        "tags": [
          "trust centers"
        ]
      }
    },
    "/v1/trust-center/domains": {
      "post": {
        "description": "Register a hostname and return its DNS ownership challenge.",
        "operationId": "createTrustDomain",
        "parameters": [
          {
            "in": "header",
            "name": "X-Watchtower-Organization",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Watchtower-Organization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Watchtower-Actor",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Watchtower-Actor",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrustDomainCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustDomainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register a custom trust center domain",
        "tags": [
          "trust centers"
        ]
      }
    },
    "/v1/trust-center/domains/{domain_id}:activate": {
      "post": {
        "description": "Require the verified hostname to point directly at the configured edge.",
        "operationId": "activateTrustDomain",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Watchtower-Organization",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Watchtower-Organization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Watchtower-Actor",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Watchtower-Actor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustDomainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authorize custom-domain TLS provisioning",
        "tags": [
          "trust centers"
        ]
      }
    },
    "/v1/trust-center/domains/{domain_id}:disable": {
      "post": {
        "description": "Immediately remove a hostname from public routing and TLS authorization.",
        "operationId": "disableTrustDomain",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Watchtower-Organization",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Watchtower-Organization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Watchtower-Actor",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Watchtower-Actor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustDomainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable a custom trust center domain",
        "tags": [
          "trust centers"
        ]
      }
    },
...
  },
...
}